factored out device info enums
